+++ user-biz/src/main/java/com/jaagro/user/biz/mapper/CustomerUserMapper.java
@@ -1,42 +1,42 @@
 package com.jaagro.user.biz.mapper;
 
-import com.jaagro.user.api.dto.response.UserInfo;
+import com.jaagro.constant.UserInfo;
 import com.jaagro.user.biz.entity.CustomerUser;
 
 public interface CustomerUserMapper {
     /**
      *
-     * @mbggenerated 2018-08-18
+     * @mbggenerated 2018-08-23
      */
-    int deleteByPrimaryKey(Long id);
+    int deleteByPrimaryKey(Integer id);
 
     /**
      *
-     * @mbggenerated 2018-08-18
+     * @mbggenerated 2018-08-23
      */
     int insert(CustomerUser record);
 
     /**
      *
-     * @mbggenerated 2018-08-18
+     * @mbggenerated 2018-08-23
      */
     int insertSelective(CustomerUser record);
 
     /**
      *
-     * @mbggenerated 2018-08-18
+     * @mbggenerated 2018-08-23
      */
-    CustomerUser selectByPrimaryKey(Long id);
+    CustomerUser selectByPrimaryKey(Integer id);
 
     /**
      *
-     * @mbggenerated 2018-08-18
+     * @mbggenerated 2018-08-23
      */
     int updateByPrimaryKeySelective(CustomerUser record);
 
     /**
      *
-     * @mbggenerated 2018-08-18
+     * @mbggenerated 2018-08-23
      */
     int updateByPrimaryKey(CustomerUser record);
 
@@ -59,5 +59,5 @@ public interface CustomerUserMapper {
      * @param id
      * @return
      */
-    UserInfo getUserInfoById(Long id);
+    UserInfo getUserInfoById(Integer id);
 }
